% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LimDist.R
\name{LimDist}
\alias{LimDist}
\title{Tool to computate the limiting distribution for a Continuous Time Markov Chain, CTMC.}
\usage{
LimDist(X, rate)
}
\arguments{
\item{X}{matrix, represents a rate matrix of a CTMC or the transition probability matrix of the DTMC associated to the CTMC.}

\item{rate}{boolean, if rate is equal to FALSE then the argument X represents the transition probability matrix of the DTMC associated to the CTMC.}
}
\description{
\code{LimDist} is used to obtain the limiting distribution of a homogeneous continuous time Markov chain.
}
\references{
Ross, S, Introduction to Probability Models, Eleven Edition. Academic Press, 2014.

Kulkarni V, Introduction to modeling and analysis of stochastic systems. Second Edition. Springer-Verlag, 2011.
}
\author{
Carlos Alberto Cardozo Delgado <cardozorpackages@gmail.com>.
}
