% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.modesto}
\alias{summary.modesto}
\title{summary.modesto}
\usage{
\method{summary}{modesto}(object, ...)
}
\arguments{
\item{object}{an object of the class modesto. This object is returned from the call to LimDist() function.}

\item{...}{other arguments.}
}
\description{
summary.modesto displays the summary of the calculated quantities from an object of class 'modesto'.
}
